requirements.txt
flask
Flask-Bootstrap4
flask-wtf
flask-testing
